+++ packages/client/components/ScopePhaseAreaGitLabScoping.tsx
@@ -1,0 +1,31 @@
+import graphql from 'babel-plugin-relay/macro'
+import React from 'react'
+import {useFragment} from 'react-relay'
+import {ScopePhaseAreaGitLabScoping_meeting$key} from '../__generated__/ScopePhaseAreaGitLabScoping_meeting.graphql'
+// import GitLabScopingSearchBar from './GitLabScopingSearchBar'
+import GitLabScopingSearchResultsRoot from './GitLabScopingSearchResultsRoot'
+
+interface Props {
+  meetingRef: ScopePhaseAreaGitLabScoping_meeting$key
+}
+
+const ScopePhaseAreaGitLabScoping = (props: Props) => {
+  const {meetingRef} = props
+  const meeting = useFragment(
+    graphql`
+      fragment ScopePhaseAreaGitLabScoping_meeting on PokerMeeting {
+        # ...GitLabScopingSearchBar_meeting
+        ...GitLabScopingSearchResultsRoot_meeting
+      }
+    `,
+    meetingRef
+  )
+  return (
+    <>
+      {/* <GitLabScopingSearchBar meeting={meeting} /> */}
+      <GitLabScopingSearchResultsRoot meetingRef={meeting} />
+    </>
+  )
+}
+
+export default ScopePhaseAreaGitLabScoping